The Covid19 Prediction Module
The Repository contains three file the module, the dataset and the final results. The data set Covid19india is containing data of the patients of different states that has firstly been detected as Covid Positive. the dataset has to first clean and we have to drop out the false and non usable data and we have splitted data to predict the results through the train data. The Covid 19 predictions.csv file is the result of the module.